+++ tests/unit/Codeception/Module/AssertsTest.php
@@ -7,7 +7,7 @@ class AssertsTest extends PHPUnit_Framework_TestCase
         $module->assertEquals(1,1);
         $module->assertContains(1,[1,2]);
         $module->assertSame(1,1);
-        $module->assertNotSame(1,true);
+        $module->assertNotSame(1, '1');
         $module->assertRegExp('/^[\d]$/','1');
         $module->assertNotRegExp('/^[a-z]$/','1');
         $module->assertEmpty([]);
@@ -20,6 +20,24 @@ class AssertsTest extends PHPUnit_Framework_TestCase
         $module->assertFalse(false);
         $module->assertFileExists(__FILE__);
         $module->assertFileNotExists(__FILE__ . '.notExist');
+        $module->assertInstanceOf('Exception', new Exception());
+        $module->assertInternalType('integer', 5);
+        $module->assertArrayHasKey('one', ['one' => 1, 'two' => 2]);
+    }
+
+    public function testExceptions()
+    {
+        $module = new \Codeception\Module\Asserts(make_container());
+        $module->expectException('Exception', function() { throw new Exception; });
+        $module->expectException(new Exception('here'), function() { throw new Exception('here'); });
+        $module->expectException(new Exception('here', 200), function() { throw new Exception('here', 200); });
+    }
+
+    public function testExceptionFails()
+    {
+        $module = new \Codeception\Module\Asserts(make_container());
+        $this->expectException('PHPUnit_Framework_AssertionFailedError');
+        $module->expectException(new Exception('here', 200), function() { throw new Exception('here', 2); });
     }
 
 }